Add files via upload
This was the code I created for Section 8 of the IBM Data Analyst Certificate program. Once the code was generated and I ran the application in the Skills Network Labs the ability to see different charts and graphs per year and per section was generated. I will be including 4 snap shots of that as well.